Split new module tests and add slow markers
