+++ pytest.ini
@@ -1,3 +1,5 @@
 [pytest]
-markers = smoke
-python_files = tests/test_*.py
+markers =
+    smoke
+    slow: marks tests as slow (deselect with '-m "not slow"')
+python_files = tests/test_*.py tests/slow/test_*.py
